add constraint enforcing only one of alt_text/text being set
we show alt text if there's no text. if there is text for an email
branding option, we show that to the right of the logo and then there's
no need for alt text at all

enforce this with a database constraint so that we can be sure all our
email_branding options are accessible. Make the constraint specifically
expect exactly one to be null to ensure we don't get both being set,
which would be confusing to reason about what is being shown

catch the constraint exception and return a nice 400 if it does happen