Generate daily report for new email brands
Every (week)day, generate a report and log it into Zendesk listing all
of the new email brands created since the last weekday so that they can
be manually reviewed for quality. We may want to tweak them or remove
them from an organisation's branding pool if it doesn't make sense for
the brand to be more widely available.