Add test for sending email with brand
Utils v58.0.0 changed the name of a parameter passed into
HTMLEmailTemplate from brand_name to brand_alt_text, which requires code
changes to apps using it.

The API uses utils, and uses HTMLEmailTemplate, but doesn't actually
have any tests covering the affected code because all of the services
used to test sending emails didn't have a branding.

This adds a test to make sure that emails can be sent whether they have
a brand or not. This test currently fails and will do until code changes
from a separate (existing) PR are merged and this is rebased.