add an endpoint to generate a name for a provided alt text
alt text is not unique, it doesn't need to be. however, names should be
unique so people can distinguish email branding options in a list of
them.

this endpoint takes in a prospective alt_text, and checks the DB. if
there's nothing in the DB with that name, it returns the alt text as
provided.

If there is, it tries `{name} (alternate 1)`, and goes up by 1 each
time. It compares case-insensitive, though always returns a name with
the same case as the input alt_text.

If a branding has 100 alternates already, it gives up. We should
probably look into what's happening before that.