Persist reference provided by services in CSV
When a service sends notifications by uploading a spreadsheet with
recipients and personalisation, so through sending a job,
if that service provides a column called reference, we will persist
the value of this column to the "client_reference" field for that
notification entry in our db.

This can help services identify their notifications easier
in the weekly reports they download. It can also help them
identify the notifications better if we provide them the information
from notification_history table.