Allow all scheduled jobs to be cancelled
The function used to find a job to cancel checks two things:

1) That it's in the SCHEDULED state
2) That it's scheduled for the future.

This means that a job which is scheduled to send, but past the scheduled
time, cannot be cancelled. But there are cases where users may want to
cancel these jobs still. We know that a job hasn't started processing if
the state is SCHEDULED, so it should still be safe to cancel even if the
ideal processing start time has passed. Let's remove the check on
scheduled_for and only require that a job's state is SCHEDULED to
determine if it's cancellable.