Add a weekly report for low inbound SMS numbers
We're rolling out some changes soon to allow services to enable inbound
SMS by themselves. At the moment they have to contact support and we'll
manually give them a number, which means we are constantly aware of how
many numbers we have left and can spot when we might need more.

If users can take from the pool themselves, without any oversight, we
need some monitoring in place to alert us automatically that we need to
ask for more numbers.

We don't know how quickly numbers will get used, but expect that a
weekly report should be fine, and that 50 numbers are unlikely to be
used up within the (roughly) week that it takes for new numbers to be
given to us.