Add a GitHub Action check to block fixup commits
It can be useful/easy to use fixup commits to make changes to a PR based
on review comments. They make it easy for a reviewer to see and only
re-review what's changed since they last commented, but we don't want
them to be merged to main.

This GitHub action will add a status check to the PR that can block
merge if any fixup commits are present, forcing the dev to squash them
down sensibly to ensure a cleaner git history.