rename NOTIFY_ENVIRONMENT from live to production
This includes renaming NOTIFY_ENVIRONMENT, which is used in the statsd
repo to prefix stats [1]. However, prometheus then renames live to
production anyway [2], so this change is safe to make.

There's two very similarly named variables here. The OS environment
variable NOTIFY_ENVIRONMENT - this is set in the makefile to be the paas
space[3] (so "production"), and then is used to look up the config in
the dictionary. The old `Live` config then manually set the config var
current_app.config['NOTIFY_ENVIRONMENT'] to `live`. This variable is
used in a bunch of places to check if we're on prod or not.

Change all these places to refer to "production" and "production" only.

[1]: https://github.com/alphagov/notifications-utils/blob/8c8c8fb591e27f11fab5cf17c64aead0870d40d2/notifications_utils/clients/statsd/statsd_client.py#L54
[2]: https://github.com/alphagov/notifications-cf-monitoring/blob/c9dbd679403fc65615f201f21833a5e2470e7f1e/prometheus/templates/prometheus.yml.tmpl#L45-L49
[3]: https://github.com/alphagov/notifications-api/blob/828d7546946d6776bf1568e38ee42ce11199f08d/Makefile#L121