reinitialise flask app after setting the database URI
flask-sqlalchemy now creates the engine when it's initialised, rather
than when the engine is first used.

this means that it tries to connect to `test_notification_api`, which is
defined in the config, rather than the worker-specific db we create in
the `_notify_db` fixture. Lets just reinitialise the db app to ensure it
gets rid of the old db name.