Fetch billing data for services in bulk
When fetching usage data for an organisation, we get the billing data
for each service individually. If the organisation has a lot (say, 100)
services, we make 100 DB calls, which is very inefficient.

Let's update `fetch_billing_data_for_day` to take a list of service IDs,
and retrieve billing data for all of those service IDs in one go rather
than having to make a query for each service.

The only saving here is that we make a single call to retrieve Services
from the DB, rather than one call per Service to retrieve. We only use
the `id` from the service, so we'll optimise this again in a bit.

We also update the `delete_billing_data_for_service_for_day` command to
use the same optimisation.

In my example case, this reduced queries on the org dashboard view from
1305 to 1206. Still some way to go ...