Fetch billing data for services in bulk
When we fetch billing data, we do it with 1 query per service, per
notification type, so if we have 100 services we're making up to 300
queries.

Let's fetch billing data for all services in one go, so that we are
making (generally always) just 3 queries.

We move the service permissions check into the query rather than the
python code to enforce that check, since we no longer have the services
as python objects.

We also don't need to retrieve the entire Service object - all we need
is the ID, and so let's collect a list of service IDs and just pass that
through instead.

From my test case on the organisation dashboard view, this reduced
us from 1206 DB queries to 908. Still some way to go ...