Add a test to limit services-with-usage queries
Let's add a test that asserts the number of queries executed by the
`/services-with-usage` endpoint, so that we don't get back to a place
where the number of queries executed by the view increases based on some
factor (eg the number of services in the org). This should help prevent
regressions in performance on this view.