Ignore prompt_toolkit module for freezegun
Freezegun has an edge case interaction with `prompt_toolkit` that ends
up breaking commands in ipdb (typed text doesn't appear on the prompt,
which makes it really hard to use).

There's a PR on freezegun to ignore this module, which would mean we
don't need to manually configure it, but for now it's not merged.