Add new go live columns to model
We need to add the columns to the model so that they can be written and
read by the REST endpoints.

By setting a default value we know that any new services or
organisations will have a value for these fields while we are migrating
them to be nonnullable in subsequent pull requests.