Prepare for per-notification-type rate limits
Refactor the existing rate limit code to understand the concept of a
notification type, which will make it easier to start enforcing the new
rate limits based on per-type usage.

This patch does *not* enforce any of the new rate limits and should not
entail any functional changes whatsoever. It only lays the groundwork.

Refactors the `check_service_over_daily_message_limit` slightly to:

1) 'loop' over the rate limits to enforce - currently just the all
notifications 'bucket'.
2) Not return any values. This was only used by the `process_job` celery
task in order to do some additional checks that the job won't go over
the daily limit for all of the notifications it wants to send. We move
support for these checks into this function, and annotate the raised
exception with information that might be needed for the failure case.