Add endpoint to notify org of go live request
This mostly copies the approach taken to Notify a service’s team members
when a service goes live.

I’ve chosen to put this in the organisation blueprint because it’s
ultimately emailing the team members for the whole organisation, even if
it’s just about one service.

It also assumes that the admin app has done appropriate checks to make
sure that the service is ready to go live, that it has an organisation,
that the organisation can approve its own go live requests, etc.