Change slow threshold for sms provider to 15% slower than 5 mins
We've had two incidents in the past few months where our users
noticed slow delivery but our load balancing algorithm didn't mark
either provider as slow.

We change the cutoff time from 4 mins to 5 mins to make it a 'nicer'
number. 4 is a slightly weird one and 5 feels rounder. This will
become more important in a future commit that will put in measurements
for delivery slower than 1 minute, 5 minutes and 10 minutes.

For the threshold, we halve it to 15%. The choice of 15% is a little
bit arbitrary. We want it to be less than 30% as this has not
identified slow delivery but our users have in the past few months.

A threshold of 15% would likely have caught the incident on 6 Dec where
approx 17% of messages were taking longer than 4 minutes to deliver
for one of our providers for a susatined period of time.

We could consider lowering this threshold even further in the past
but a downside of lowering it too much is it incentives our sms
providers to fail quickly rather than take a while to deliver a
message that genuinely needs over 5 minutes to be delivered (say
due to a handset being off).

Let us see how we get on with this new threshold and evaluate again
if we want to change it.