Move calculation for slow delivery out of `is_delivery_slow_for_providers`
This is so the functionality can be reused for calculating delivery
stats where we don't determine if a provider is slow or not

I also try to give better names to some of the variables in this
commit but open to discussion on whether it is better or not.

I also move the timedelta code into
`get_ratio_of_messages_delivered_slowly_per_provider`
so that now that we make a few calls to that function, it is a bit
easier to call without needing to deal with timedeltas.