Initialise DVLA API client and get fresh JWT token
Create an initial version of the DVLA API client which knows how to
authenticate and retrieve a JWT token that is valid for 1 hour.

This is mostly boilerplate client code for sending requests. The main
difference from our other clients is the need to pull credentials from
AWS SSM, and have short-lived (JWT) auth tokens that need rotating
within an app's lifetime. Because the credentials need to change
regularly, it's not practical to re-deploy the API whenever a credential
changes, so we can't provide them via environment variables.

We have four main pieces of data that are used to authenticate to DVLA:
1) username
2) password
3) api key
4) JWT

The username is a fully-static string.

The password is secret that has a lifetime of 90 days and then needs to
be rotated.

The api key is a secret that has a lifetime of 1 year and then needs to
be rotated.

A JWT token can be requested from the DVLA Authentication API, and has a
lifetime of 1 hour. You need to provide the username and password to get
a valid JWT token.

When the password is rotated, any active JWT tokens will continue to
work until they expire automatically after an hour.

When the API key is rotated, JWT tokens will immediately* become invalid
and need to be rotated.

* within a few seconds

We add a `get_print_job` method that hits the DVLA Letters API, as this
endpoint gives us a 401/403 on auth errors, and a 404 on
successfully-authenticated requests. So we can use this endpoint to
check whether or not we're authenticated.

Co-authored-by: Sam Williams <sam.williams@digital.cabinet-office.gov.uk>