Rotate API key
As for rotating the password, now with the DVLA API Key.

When rotating the API key, though, existing JWT tokens *are* immediately
invalidated (note: within a few seconds). We re-use the existing
credentials lock so that only one client can be changing the password or
the api key at any time.

The DVLA API hands us a new API key - we don't tell it what that key
should be (unlike with passwords).

Co-authored-by: Sam Williams <sam.williams@digital.cabinet-office.gov.uk>