Rotate password
Adds a task and a DVLAClient method to rotate the password using the
DVLA Auth API.

We generate a secure ~32 character password with characters from all of
the main character classes, as DVLA has some strict password structure
requirements.

When we successfully change the password, the client also handles
storing that updated password in AWS SSM. It does not need to
automatically get a new JWT, as rotating the password does not inactivate
existing JWT tokens.

We add a redis lock around the password rotation logic so that, in
theory, two clients cannot change the password at the same time and
cause some race condition issues.

Co-authored-by: Sam Williams <sam.williams@digital.cabinet-office.gov.uk>