Retry logic
There are three credentials that can expire on us at any time:
1) JWT token (lives an hour)
2) Password (lives 90 days)
3) API Key (lives 1 year)

Our client should be able to automatically recover from any of these
credentials expiring during the app's (api's) lifetime. We should expect
to use JWTs right up until they expire (eg for an hour). However, we
should have processes in place that rotate the password and API key well
in advance of their theoretical lifetime so that we've always got some
room for error and for manual intervention. Perhaps we could rotate the
password every week, and the API key every month.

When requests to the DVLA API fails due to expired credentials, we can
get various error codes back.

A 401 generally indicates that the JWT has expired. This is the most
likely auth error for us to hit, so whenever we hit a 401 let's
automatically try refreshing the JWT and trying the request again.

If we are unable to get a new JWT token and receive a 40x, it will be
because the password is no longer valid. In this case, we believe that
our automatic process should have resulted in a new password being
issued, so let's fetch the new password from AWS SSM. Then we can try
one more time to get a JWT token. If this fails, we don't know why,
because we've got the latest password. So in this case, kill the worker
and let is start up.

If we do authenticate correctly the 2nd time and receive a JWT, we know
the password is correct. Let's then retry the original request. If
_that_ request fails with a 40x, then the JWT is invalid because it's
associated with an old API key. So in this case the answer is also to
refresh credentials from AWS SSM, because the API key has been rotated
by our automated process. We then authenticate and try the request one
more time - if this fails, as with the password, we kill the worker
because we're unable to recover automatically.

In `dvla_tasks.py` we add a few test flows that will fire off a load of
tasks which should automatically rotate the credentials, validating that
the client can recover from various authentication issues automatically.

We have the following flows:
1) Check that we can handle expired JWT tokens
2) Check that one client can rotate the password and another can detect
that, pull the new password, generate a new JWT, and continue without
issue.
3) Check that one client can rotate the API key, and another can detect
that, pull the new API key, generate a new JWT, and continue without
issue.

Co-authored-by: Sam Williams <sam.williams@digital.cabinet-office.gov.uk>