Record when notification statuses change
Start recording key events in a notification's lifecycle. The main one
here is when a notification enters a new status (eg created, sending,
failed, delivered).

We also record when the provider who will deliver the notification is
set, as well as when duplicate callbacks are received. This gives us one
place to look for all the important events that have happened to a
notification.

We only retain this detailed log of events while the notification is in
the main notifications table. When they are migrated to
notifications_history, we delete this record. This means that we retain
the information according to each service's data retention policy, which
is 7 days by default but can be extended on request.

Add endpoint to retrieve notification events

We only want to be able to retrieve notification events as an admin
user, currently, so we add an endpoint here to the services blueprint,
in line with the current admin get notification request, rather than to
the public notifications endpoints.

Remove unnecessary join onto Notifications

Record events when updating notifications by reference

Record events when timing notifications out

Add note for nightly timeout event

Add event for cancelled letter jobs

Reformat duplicate callback notes

Remove sent_by (provider) as this is already recorded in a previous
event