Fix edge case for bulk sending
There is a theoretical edge case for the new rate limits when posting
bulk jobs.

If the first notifications sent by a service on a given day are a bulk
job, and that bulk job exceeds the service's limits, we don't fail the
job in advance.

This is unlikely to have ever been hit in reality, as bulk jobs are
limited to 100k and our new per-channel limits are going to be 250k, but
we should fix this anyway to 1) prevent confusion and 2) not cause any
future issues if we dropped the limits in the future.