Migrate per-channel rate limits
We want to start applying the new per-channel rate limits soon. This
migration will set the limits at the levels we want to apply:

For services in trial mode, 50 for each notification channel.

For live services, 250k for emails and SMS, and 20k for letters.

We only overwrite the limits if they're at the current default
(insanely high) value, which means that any rate limits we've set
manually for specific services won't be affected.