Add new columns for letter branding
This adds new columns to the letter branding table to record when a row
was created or updated and who by. The email branding table uses
slightly different column names (`created_by` and `updated_by` instead
of `created_by_id` and `updated_by_id`). Adding `_id` on the end of the
column names is more consistent with the other tables, so we can
potentially change the email branding column names later.