remove research mode
we haven't used it for years, and it's just confusing and takes up lines
of code and adds complexity.

unfortunately we've still got the "research" worker and
"research-mode-tasks", these only correspond to test api keys now so it
would be nice to rename them but that's a lot trickier to do.

removing the column definition means we no longer read or write - the
column is now nullable so this is okay. a future PR will drop the column
entirely