Update Python rate.rate field type
The `rate` column of the `rates` table is a NUMERIC in the DB, but was
defined as `db.Float` in Python, which meant alembic kept wanting to
change it in a migration.

Unlike migrating text types, this feels slightly more risky in that it
could affect rounding on numbers. So let's just update the Python
representation. We have other numeric fields, so it's not like we are
particularly consistent.