Make services_history.prefix_sms non-null
This model is dynamically generated from the Service model, so it's not
easy to have separate definitions for the columns on the main and
history model.

The main Service model has prefix_sms as non-null, which means they all
have values. We can probably accept overwriting the historical
prefix_sms nulls.