Change evaluation period for slow sms from 10 mins to 15 mins
In the previous commit we added metrics for the percentage of messages
sent in the last 10 minutes that have taken over 1,5 and 10 minutes
to be delivered.

Pea spotted that the 10 minute metric would not work because you can't
have a message created less than 10 minutes ago having taken more than
10 minutes to deliver.

We think this 10 minute delivery metric is valuable and want to make
it work. Our suggested way of doing this is to extend the measurement
period from messages sent in the last 10 minutes to the last 15 minutes.

I think it doesn't make too much sense to have all the other metrics
looking at the last 10 minutes of messages sent, but this one looking
at 15 minutes so think the best thing to do is to have all of our
slow delivery metrics consider messages sent in the last 15 minutes.