don't update returned letters
previously, if a returned letter was referenced in two separate DVLA
reports, we would update the existing row in the database - updating the
`reported_at` date and setting the `updated_at` timestamp.

However, this leads to confusing behaviour - the user might have
downloaded an old report, then when they check again, that report will
have _less_ rows, and the new report will have rows that they may have
already handled and actioned a month prior.

Change the code so that if the row already exists, it just skips the
insert.