Stop sending stuff to the priority queue
Originally we had the idea that we could set certain templates (for
example two factor codes) as ‘priority’ and try to deliver them more
quickly.

We only ever let platform admins decide whether a template was priority.

Since then we discovered that:
- Celery 3 was slower to deliver messages from an often-empty queue
  (which the priority queue was) so this feature had the opposite effect to what was intended
- Celery 4 is so fast the it obviates the need for a priority queue
- We have so many templates that it’s beyond the capacity of our team to
  review which ones should or shouldn’t be ‘priority’

By not sending any notifications to the priority queue, we can safely
remove (in a followup PR):
- the queue
- the column on the templates table and model

This will affect about 10 templates in production which belong to real,
live services.