Update `rebuild_ft_billing_for_day` command
This command no longer matched how we populate the ft_billing in the
overnight task. If no service_id was provided, it would use the
`get_service_ids_that_need_billing_populated` function to find out which
data to populate the table with. However, this only looks in
notification_history, and we now look in the notifications table too.

This changes the dao function that is used to delete the billing data
for a day so that the service_ids argument is optional, which means we
no longer need to use `get_service_ids_that_need_billing_populated`.