use redis locks to prevent two processes updating dvla creds at once
We don't want two processes to update the creds. We could end up in a
situation where one process updates the password, then another process
updates the password again, saves it in SSM, and then first process stores the (now
irrelevant) first password in SSM, and we lose the actual password that
is saved in DVLA.

we acquire a lock when making the redis calls. The lock is saved in
redis, and accessed through redis-py's built-in Lock class. We call it
with a timeout of 60s (such that if something goes wrong, we do release
the lock and let another process try), and we call it with
`blocking=False`, so if a second process tries to access it it'll error
rather than wait. We don't need to wait and then set the thing multiple
times, we can just accept that another process is already handling this,
and we can then wait a few seconds and check SSM for the new creds.

raises a redis.exception.LockError if another process is currently
setting the value.

There's a separate lock for each credential and each DVLA username