Add error handling to `send_letter` method
This checks the status code that is returned by the DVLA api when we
send a letter and raises a custom exception based on the code returned.

The exceptions will be handled later, but for now we just need to know
what type of error it was and whether it is one that we can retry. Where
an error has details, we show these as part of the error message, for
example if there are details about why the request was invalid.