create scheduled tasks to reset DVLA password and api key
our dvla password needs to be rotated every 90 days
our api key needs to be rotated every year

if the password expires or is changed without the creds being updated
locally we won't get downtime, existing jwt tokens generated with the
old password will continue to work, however, we won't be able to
generate new tokens until the password has been reset and the new
password has been retrieved from the creds.

if the api key expires (or is changed by another process) then all
requests will fail until we regenerate the api key and get the new one
from the creds.

to avoid either of these situations occurring unexpectedly, we'll reset
the api key on the first tuesday of every month and the password on the
following wednesday. we do this so that if anything does go wrong, it's
likely to be a workday (tuesday and wednesday so they're less likely to
be affected by bank holidays, eg new years day, may day bank holiday).
we run them on separate days so there's less likely to be issues if
something goes wrong - especially since we need an up to date password
to reset the api key. the password reset will always happen the day
after the first tuesday, even if the first wednesday was the week
before. Run it at 9am so it's unlikely to affect any DVLA letter
sending, but won't have too long to wait before we log in and debug if
there are any errors or issues

the test is pretty boring right now but we may add error handling, retry
logic etc