handle auth errors when calling endpoints.
a key point to note is that the print API returns the json
`{"errors": [...]}`, while the change password, api key, authenticate
endpoints return just the list `[...]`.

The contents of these lists are inconsistent, they're always dicts with
keys `status` (sometimes an http code) and `detail` (a sentence describing the
problem), but otherwise they're different. Some examples:

print-request with valid creds and empty json body. Note the different
code and message. title appears to refer to the field affected. `status`
is a full string here.
```
{'errors': [{'status': '400 BAD_REQUEST',
   'code': 'NotEmpty',
   'title': 'customParams',
   'detail': 'Custom parameters field must not be empty.'},
  {'status': '400 BAD_REQUEST',
   'code': 'NotNull',
   'title': 'standardParams',
   'detail': 'must not be null'}]}
```

print-request with invalid creds (either jwt or api key being bad). Note
the status being just the number part, and the code being the http
status's name
```
{'errors': [{'status': '401',
   'code': 'Unauthorized',
   'detail': 'API Key or JWT is either not provided, expired or invalid'}]}
```

when we call new-api-key we get different error bodies, note `status` is
an integer, we have `title` with the http status name
```
[{'status': 401,
  'title': 'Unauthorized',
  'detail': 'API Key or JWT is either not provided, expired or invalid.'}]
```

essentially I think the only thing we can rely on is the actual response
status code and the detail value in the json response.