clear invalid creds out of cache if we get a 401
general theory here:

if creds are out of date, clear them from the cache. Don't handle
resetting them at this point as we're not sure on the context under
which this error has been received (eg: are we trying to authenticate to
send a letter, vs to reset an api key). We don't particularly want 500
concurrent requests all triggering a password reset at once if they fail
to authenticate.

We don't expect to ever be using expired JWT tokens (thanks to our local
memory cache for the jwt token monitoring its expiry time through
dvla_client._jwt_expires_at. This means that if, for example, the
print-letter request fails (which takes both a jwt token and the api
key), we assume it is the API key that is an issue.

if we're making an authenticated request, that may call /authenticate
while preparing the headers. If that call throws a 401, then the
/authenticate exception handler will clear the cache and wrap that in a
DvlaRetryableException, which is _not_ re-caught by the outer exception
block for the /print-request or /new-api-key method calls. This means
that if we catch a 401 from within, eg, the change_api_key exception
hadnling block, we know it's from the new-api-key call and not from the
authenticate call.

This may be revisited in the future.