clear password/api key before trying to change them
this guards against the case where one task changes a password
succesfully, and then another password tries to change the password.
Previously it would try and use that instance's old cached value (which
would error as that password is out of date). Now it will ensure it's
got the latest value. If this still throws an error, there's something
really wrong that we need to investigate (the value in SSM is not valid
when resetting)