Use notification reference for DVLA API
In calls to the DVLA API, let's pass our notification reference through
as the businessIdentifier. This maps to the 'reference' we get back from
DVLA in the daily delivery receipt files, so if we send the reference
through the API then we should be able to match delivery receipts up
automatically without having to make any specific changes.