Proof of concept: Notification events via celery
The new event log record is causing our POST notification endpoint to
spend too much time in the DB, which leads to requests waiting on DB
connections and eventually timing out.

The notification events don't _need_ to be persisted to the DB in the
request cycle, so let's instead just put the events on a queue and
process them asynchronously (but near-real-time).