Refactor DVLA client to take in API URL
We want to restrict connections to the DVLA API to use a specific set of
SSL ciphers. In order to do this we'll need to mount a request adapter
with a prefix of the base URL. So let's take this in and refactor the
client to always use that base url when making requests.