Restrict TLS connections to specific ciphers
The DVLA API we will connect to in prod supports a number of insecure
ciphers. It will take some time for them to retire support due to other
clients requiring them, but we want to make sure we aren't able to
connect with those insecure ciphers.

We will supply our own list of acceptable cipher suites (sourced from
NCSC here:
https://www.ncsc.gov.uk/guidance/using-tls-to-protect-data#section_6)
and limit to only connecting with one of those.