Bump utils to 62.0.1
This pulls in a new `normalised_lines` format for BFPO addresses. We now
put the BFPO number on the last line, after the postcode (if present).

This is more in line with the format for BFPO addresses as recommended
on GOV.UK, so should be preferrable for our print provider/Royal Mail.
It also includes some additional utility methods for extracting BFPO
information so that we can provide normalised data to our print
provider's API to mark up BFPO addresses.