Require DB before running UserIdFilter tests
This is a bit of a hack that addresses some weirdness in our test setup.

We have a fixture `_notify_db` that does some initialisation on the
(Flask) `app` instance. We can only do that initialisation if no
requests have been made to the app yet. These tests don't require the DB
but may make requests to the app; if that happens before any other test
accesses the DB fixture then the DB fixture initialisation will fail
(because requests have already been handled).

We should rework our tests to handle this more cleanly - but for now
marking these tests as requiring the DB is fine.