set letter to technical-failure if we cant find it in s3
otherwise we'll just do nothing until the next day, when we'll try and
send it again. i think the chance of it appearing in s3 overnight feels
slim to unlikely - this was behaviour we added three years ago[^1] but
the log message never fired due to a bug in how the function was called.
it feels safer and more correct to set the letter to technical-failure.
we'll get a proper exception log and it'll be easier to distinguish
failed letters in technical-failure that we need to action from new
letters in created that will go out as normal

[^1]: https://github.com/alphagov/notifications-api/pull/2760