Filter subquery by organisation
Allow filtering `fetch_usage_for_all_services_sms` to only return
services for a specific organisation. This function powers part of the
org dashboard, which is a very heavy page. Without this filter, a
subquery is constructed that does a join on *all* services against the
ft_billing table. We only need to do a join on a specific
organisation's services - so this is really wasteful.

This reduces query execution in a fairly flat way across a variety of services. The new query runs in approximately
30% of the time.

| Organisation            | Query execution time: before (ms) | Query execution time: after (ms) |
|-------------------------|-----------------------------------|----------------------------------|
| Cabinet Office          | Time: 1904.782 ms (00:01.905)     | Time: 589.321 ms                 |
| HMRC                    | Time: 1928.833 ms (00:01.929)     | Time: 547.078 ms                 |
| DWP                     | Time: 1849.638 ms (00:01.850)     | Time: 570.952 ms                 |
| MK Council              | Time: 1837.129 ms (00:01.837)     | Time: 570.980 ms                 |
| Brighton & Hove Council | Time: 1840.022 ms (00:01.840)     | Time: 544.516 ms                 |