Add another filter on ft_billing join
We can further optimise this query by restricting the join between
services and ft_billing to filter ft_billing upfront on services in the
provided organisation. This allows the DB to take advantage of indexes
more cleanly, moving from index scans with filters to index only scans.

This reduces query execution in a variable way depending on how active the organisations's services are.

| Organisation            | Execution time: baseline | Execution time: commit 1 | Execution time: commit 2 |
|-------------------------|--------------------------|--------------------------|--------------------------|
| Cabinet Office          | Time: 1904.782 ms        | Time: 589.321 ms         | Time: 387.585 ms         |
| HMRC                    | Time: 1928.833 ms        | Time: 547.078 ms         | Time: 136.522 ms         |
| DWP                     | Time: 1849.638 ms        | Time: 570.952 ms         | Time: 204.146 ms         |
| MK Council              | Time: 1837.129 ms        | Time: 570.980 ms         | Time: 39.411 ms          |
| Brighton & Hove Council | Time: 1840.022 ms        | Time: 544.516 ms         | Time: 44.034 ms          |